This patch fixes a bug in the test/report process.
Each test is divided into several steps:
 1. startup/initialization
 2. compression
 3. comparison of ROHC packets
 4. decompression
 5. comparision of uncompressed packets
 6. shutdown/cleaning

The test program reports problems occuring during steps 2-5, but not those
occuring during steps 1 and 6. The shutdown step can not fail (memory cleaning
only), but the startup step may eventually fail (missing PCAP file for example).
If the startup fails, all the following steps are ignored and the report script
prints a PASS status for them. This is obviously wrong and can lead to bad
reports.

The patch alters the XML ouput of the test program: the startup step now outputs
its status. The report script prints this status on the HTML page and changes
the status of the other steps to FAIL when the startup step fails.