This patch fixes a CFP-related bug in the UDP-Lite decompression.
Context:

The dynamic part of the UDP-Lite header may contain the Checksum Coverage field
if necessary. It depends on the type of packet (the CC field is mandatory for
IR and IR-DYN packets) and on the CC value (the CC field is mandatory if the
value is different from the UDP-Lite packet length).

The decompressor knows if the CC field is present with the CFP flag. This flag
is part of the decompression context and is initialized upon reception of IR or
IR-DYN packets (the CC field is always present in IR and IR-DYN packets). If the
CC value of an IR/IR-DYN packet is different from the UDP-Lite packet length,
CFP = 1, otherwise CFP = 0.

In order to initialize the CFP field, the udp_lite_decode_dynamic_udp() function
must compute the length of UDP-Lite packet. This computation needs to know the
type of packet (IR, IR-DYN, UO...).

Problem:

The d_generic_decode_ir() function does not set the packet_type context variable
to PACKET_IR. The patch fixes this problem. There is no problem with the other
types of packets.