This patch slightly changes the behaviour of the context check functions of the IP-only, UDP and UDP-Lite profiles.
The context check functions check whether an IP packet belongs to a given
context or not. They currently return code 0 (= bad context) for two distinct
cases:
 - the IP packet does not match the context (different IP version or IP
   addresses for example)
 - the packet can not be compressed by the profile at all (IP version different
   from 4 or 6, fragmented IP packet for example).

Bad context is ok for case 1, but not for case 2. The 2nd case must cause a
profile change, ie. the compressor must stop finding a context in the current
profile and must switch to another profile (IP-only or Uncompressed profile).

The patch fixes the problem by returning code -1 (= error code) in the 2nd case.