This patch fixes a bug on the statistics.
At the end of one packet decompression, the library stores the length of the
compressed ROHC header and the length of the decompressed IP header for
statistical purposes. The length of the ROHC header was incorrectly computed.
The first bytes (add-CID, large CID, Profile ID, CRC fields for example) were
not taken in account.

The patch fixes the computation of the ROHC header length. The prototype of the
decode_ir() function changed in order to know the length of the large CID field
and whether the add-CID field is present.