This patch fixes a bug in the test application. The problem comes from the padding in the Ethernet frame.
Padding can be added at the end of the payload of an Ethernet frame if the
payload is too short, so that the Ethernet frame is at least 60-byte long.

Example: a 44-byte IPv4 packet packed in an Ethernet frame (header = 14 bytes)
makes a 58-byte frame, leading in the insertion of 2 bytes of padding.

Now, the test application reads the IP total length in order to draw aside the
padding bytes.