Initial import of a new application that creates ROHC over UDP tunnels. This application can help in debugging the ROHC library.
Description:

The program creates a ROHC tunnel over UDP. A ROHC tunnel compresses the IP
packets it receives from a virtual network interface and decompresses the
ROHC packets it receives from one UDP flow.

               +-----------+                          +----------+
 IP packets    |  Virtual  |     +--------------+     |          |
 sent by   --> | interface | --> |  Compressor  | --> |   ROHC   |
 the host      |   (TUN)   |     +--------------+     |  packets |
               |           |                          |   over   |
 IP packets    |           |     +--------------+     | UDP flow |
 received  <-- |           | <-- | Decompressor | <-- |          |
 from the      |           |     +--------------+     |          |
 tunnel        +-----------+                          +----------+

The program outputs messages from the tunnel application on stderr and
messages from the ROHC library on stdout.

Usage:

rohctunnel NAME remote RADDR local LADDR port PORT

NAME    the name of the tunnel
RADDR   the IP address of the remote host
LADDR   the IP address of the local host
PORT    the UDP port to use (local and remote)

Example:

Type as root on machine A:

 # rohctunnel rohc0 remote 192.168.0.20 local 192.168.0.21 port 5000
 # ip link set rohc0 up
 # ip -4 addr add 10.0.0.1/24 dev rohc0
 # ip -6 addr add 2001:eeee::1/64 dev rohc0

Type as root on machine B:

 # rohctunnel rohc0 remote 192.168.0.20 local 192.168.0.21 port 5000
 # ip link set rohc0 up
 # ip -4 addr add 10.0.0.1/24 dev rohc0
 # ip -6 addr add 2001:eeee::1/64 dev rohc0

Then:

 $ ping 10.0.0.1
 $ ping6 2001:eeee::1