This patch fixes a bug in the IP-only profile.
When the compressor compresses IP packets carrying different transport protocols
(ICMP and TCP for example), the STATIC Protocol field changes and thus needs to
be sent to the decompressor (so that it can update its decompression context).

The STATIC Protocol field cannot be sent in an IR-DYN packet (no static fields).
Unfortunately, the compressor prefers sending an IR-DYN packet rather than an
UOR-2 packet (with static fields).

The patch changes the algorithm that selects the packet to send: when STATIC
fields are modified, an UOR-2 packet is sent rather than an IR-DYN packet.

The patch also adds a new test in the report. It tests the compressor with mixed
IPv4/ICMP and IPv4/TCP packets.