This patch improves the Stream Classifier.
If one or more IPv4 + IPv4 packet has already been classified (and a context
created for the stream), one IPv4-only packet may currently be classified in
the same stream (and the same context will be used for the compression). If the
IPv4 packet is received before the IPv4 + IPv4 packets, the problem does not
occur. The problem also occurs with IPv4 and IPv4 + IPv6 packets, with IPv6 and
IPv6 + IPv4 packets or with IPv6 and IPv6 + IPv6 packets.

The ROHC compression scheme is more efficient if packets with double IP headers
and packets with single IP headers are seperated in two different streams. The
patch fixes this classification problem.