This patch adds a new feedback related function. Feedback packets are currently sent only by piggybacking. This is a problem if there is very few traffic on the channel used for feedback. Feedback packets are queued and sent late to the compressor.
This patch introduces a new functionality: feedback-only packets. Feedback-only
packets are ROHC packets that do not contain any compressed data, but only
concatened feedback packets. The decompressor forwards the feedback packets to
the associated compressor and then discards the ROHC packet. Feedback-only
packets are described in the ROHC RFC.

The patch updates the library API. The rohc_feedback_flush() function allows 
applications to flush the feedback packets waiting in the queue. This function
could be called on a regular basis by applications to send feedback even if
there is no traffic.