With this patch, the tunnel application is able to emulate:  - both unidirectional and bidirectional links  - a lossy link (error rate is specified by the BER given by user)
When unidirectional mode is choosen, the decompressor is not linked with the
compressor. Thus, no feedback is transmitted. The ROHC compressor stays in
U-mode.

When a binary error hits a ROHC packet, the packet is fully dropped. The
decompressor never receives it. This can cause error propagation in next
packets.

The statistics output is modified. The packet ID is not the number of packets
received or sent, but the ROHC Sequence Number (SN). It helps drawing graphs
with error propagation.