Improve non-regression test
Add the option --rohc-size-ouput to the non-regression application in order
to write the sizes of every generated ROHC packets to a plain text file.
The plain text files are then stored in the repository along the reference
ROHC captures.

If a problem is fixed in the ROHC library and if the reference ROHC captures
must be re-generated (because the type of packet or the format of packets
changed for example), then the plain text files that contain the sizes of the
generated ROHC packets might avoid some regressions that make too many ROHC
packets larger than before.