Fix bug #555695 entitled "GCC warnings on x86-64".
The change is more invasive than for the 1.2.x and 1.3.x branches. I decided
to get rid of arithmetic on pointers. Such things are likely to be not portable
accross platforms (16 bits, 32 bits, 64 bits...).

To achieve that, I added new local variables to count the numbers of bytes read
in the packet. This required some changes in prototypes of several internal
functions. Care was taken that the result of non-regression tests was unchanged:
both ROHC packets and debug output.