Fix bug #622402: Incorrect format for UO-1-ID packet https://bugs.launchpad.net/rohc/+bug/622402
In the UO-1-ID packet format, the first bit of the 2nd byte is a X bit not
a M flag. RFC 3095, section 5.7.3, page 81 tells so. The X bit indicates
whether an extension is present, not whether the RTP Marker field changed.

Update the compressor logic to replace the M bit by a X bit (always 0 at the
moment). Also update the decompressor logic to read a X bit instead of the M
bit (drop the packet if the X bit is set to 1 at the moment).

Extention support for UO-1-ID packets shall be written later.