Allocate the array of decompression contexts to its full size to avoid re-allocations during packet processing. This is both for performance reasons and code simplicity.
Introduce the new rohc_decomp_set_cid_type() and rohc_decomp_set_max_cid()
functions to configure the CID type and the MAX_CID value that are allowed
for the decompressor.

Set the rohc_decompress_both() function as deprecated. It still works as
before, but the 2 above functions are preferred.