Improve RTP profile detection.
Add a user callback to determine if an incoming packet is a RTP packet.
Add a user API for the callback and for adding/removing/reseting UDP ports
dedicated to RTP traffic.

Add check_profile() routines in profiles. The check_profile() functions help
finding the best compression profile for a packet. The checks for the dedicated
UDP ports are now done in the check_profile() function of the RTP profile making
the generic module more generic.

Initial patch by Julien Bernard, Didier Barvaux, and Audric Schiltknecht from
Viveris Technologies. The patch was updated by Didier Barvaux for the main branch.