Fix bug #1080055: Decompression failure caused by RTP stream with lost packet but irregular IP-ID https://bugs.launchpad.net/rohc/+bug/1080055
Add a non-regression test for a RTP stream with a lost packet (SN = 8) and the
next packet that contains an irregular IP-ID:
  ...
  packet #7: SN = 7, TS = 350, IP-ID = 7
  packet #8 is lost, should be SN = 8, TS = 400, IP-ID = 8
  packet #9: SN = 9, TS = 450, IP-ID = 8

The IP-ID of packet #9 should be 9 instead of 8. The compressor chooses packet
UO-1-ID instead of UOR-2-ID. The decompressor then fails to decompress the
packet.

The problem was corrected by fixing the function that determine the number of
IP-ID bits required for inner and outer IP headers. When the stream contains
only one IP header, this outer IPv4 header is the innermost IPv4 header.