Add ROHC segment support as defined in RFC 3095 section 5.2.5.
Add 2 new API functions: rohc_compress2() and rohc_comp_get_segment().

The rohc_compress2() function replaces the now-deprecated rohc_compress()
function. It warns about required ROHC segmentation if the given output buffer
is too small or MRRU is reached. It returns the ROHC length by reference, and
a code return value. The code may be:
 - ROHC_OK if compression is successful and segmentation is not required,
 - ROHC_NEED_SEGMENT if ROHC segmentation is required (no data is returned),
 - ROHC_ERROR in case of compression problem.

The rohc_comp_get_segment() function may be used to retrieve the ROHC segments
when a call to rohc_compress2() returned ROHC_NEED_SEGMENT. It returns one
segment at a time. It returns the ROHC length by reference, and a code as return
value. The code may be:
 - ROHC_NEED_SEGMENT if a segment is retrieved and there is more to come,
 - ROHC_OK if the returned segment is the last one,
 - ROHC_ERROR in case of segmentation problem.

The rohc_decompress() function was updated to return a new code dedicated to
the processing of ROHC segments: ROHC_NON_FINAL_SEGMENT. When this code is
returned, the function does not return any decompressed data because it handled
a non-final ROHC segment. Decompressed data will be returned when the final
ROHC segment will arrive.

All the applications and test were updated to use rohc_compress2(). Some apps
and tests were updated to handle ROHC segmentation. A new test was added to
check that ROHC segmentation is working as expected.


The ROHC segment support was designed and written by Didier Barvaux for Viveris
Technologies, and Didier Barvaux on his personal time.