Fix bug #1088004: Decompression failure of specific IPv4 stream with changing RND https://bugs.launchpad.net/rohc/+bug/1088004
The ROHC library failed to decompress one IPv4 stream with changing RND
(RND=1 -> RND=0) and less than 5 bits of IP-ID delta to transmit.

As RND changed, the extension 3 is required. As there are some IP-ID bits to
transmit, the UOR-2-ID packet is chosen. As there are less than 5 bits of IP-ID
delta to transmit, the compressor puts these bits in the UOR-2-ID base header.
Because the RND flag changed, the compressor also puts 16 bits of IP-ID delta
in extension 3 (flag I=1). The decompressor fails to decompress such a packet
because it receives more than 16 bits of IP-ID delta.

Solution:
 1/ At compressor, put zeroes in the 5 bits of the UOR-2-ID base header if
    RND flag changed in the last few packets.
 2/ At decompressor, warn about > 16 IP-ID bits but do not stop parsing, simply
    ignore superfluous bits.

For the 2nd point, stop parsing if ROHC_RFC_STRICT_DECOMPRESSOR is defined.