Fix bug #1089106: Compressor asserts because of non-RTP traffic with 14 SN bits to transmit https://bugs.launchpad.net/rohc/+bug/1089106
Packet type UOR-2 with extension 3 for non-RTP profiles does not accept more
than 13 bits of SN. The decision algorithms in FO and SO states checks the
limit of 14 SN. They are wrong.

Also fix the lack of limit check when a static field changed (for both RTP
and non-RTP profiles).

Add a non-regression test for that problem.