Use the Uncompressed profile for UDP packets with a wrong length field.
UDP packets with a wrong Length field will be correctly compressed by the
IP/UDP profile, however the decompressor will not be able to re-create the
exact same UDP header since it will compute the Length field on its own.
The UDP Length field will be different from the original one. To avoid this
problem, use the Uncompressed profile for such malformed packets.