Fix bug #1092660: Problem with RTP streams with Padding bit set https://bugs.launchpad.net/rohc/+bug/1092660
If one RTP stream has the RTP Padding (P) bit set for one (or more) of its
packet, the ROHC library fails to compress/decompress it correctly if one
IR/IR-DYN packet type is not select by chance.

Extension 3 (and one of the packet type that supports extension 3) is required
to transmit the RTP Padding bit if the IR or IR-DYN packet type is not select
for another reason.