Fix bug #1093836: Decompressor crash because of lossy RTP stream with SN wraparound https://bugs.launchpad.net/rohc/+bug/1093836
When decompressing one RTP stream with SN wraparound but several lost packets
(more lost packets than the width of the W-LSB window), the decoded SN is
greater than 0xFFFF. This is not expected for RTP streams because SN is 16-bit long
max. So, the library asserts when decoding RTP TS.

The solution is to truncate the SN values to 0xFFFF max before using them.