Fix bug #1093846: Possible decompressor crash with malformed extension header list https://bugs.launchpad.net/rohc/+bug/1093846
The ROHC decompressor does not check correctly the length of the ROHC packet
before parsing the item part of the extension header list. If the item is
malformed, the ROHC decompressor may read too many bytes and parses data that
was not part of the ROHC packet. Depending on those bytes, it might fail softly
or crash.

It might have security implications if someone is able to send malformed ROHC
packets to a ROHC decompressor.

Solution: always check the ROHC packet length before reading items of extension
          header lists.