Fix bug #1092660: Problem with RTP streams with Padding bit set https://bugs.launchpad.net/rohc/+bug/1092660
Additional fix for bug #1092660. The RTP Padding bit shall be stored in the
compressor context because it is a STATIC field. It does not work as the RTP
Marker field that is set to 1 only if transmitted in the ROHC header. If the
RTP Padding bit is not present in the ROHC header, the one from the context
is used.

Enhance the bug1092660-rtp-padding-bit non-regression test to cover the case
where the Padding field is set again to 0 after being set to 1.