Fix bug #1105963: Uncompressed profile: Normal packet cannot be used for non-IPv4/v6 packets https://bugs.launchpad.net/rohc/+bug/1105963
As stated at the end of §5.10.2 in RFC 3095, the Uncompressed profile cannot
compress non-IPv4/v6 packets with Normal packets. If done, the first byte of
payload could be mis-interpreted as a different ROHC packet type (IR for
example). It would result in a decompression failure.

The test/non_regression/inputs/unknown/source.pcap non-regression test was
extended to test that kind of problem.