Fix bug #1115091: Crash when decompressing list types 1/2/3 if no reference list was already set https://bugs.launchpad.net/rohc/trunk/+bug/1115091
When (de)compressing list of IPv6 header extensions, a reference list is shared
between the compressor and decompressor. The compression types 1, 2 and 3
require a reference list to be defined. If not, the library crashes. Such a
situation might happen if some ROHC packets defining the reference list are lost
or damaged.

Solution: use the list with the received gen_id as the reference list in case
          no reference list was set yet.