TCP profile: fix handling of TCP SACK option wrt to RFC 6846.
The problem was found with the ROHC sniffer on my local network. A TCP segment
contained a SACK option with a block begin lower than the ACK number of the
previous TCP segment.

A non-regression test is added. It is the ipv4/tcp/source.pcap traffic with one
additional packet with a SACK option as described above.