Fix bug #1069411: Support Linux kernel https://bugs.launchpad.net/rohc/+bug/1069411
The Linux support comes with 2 kernel modules:
 1/ one module to export the ROHC library to the Linux kernel,
 2/ one module to use the ROHC library and test it in the Linux kernel.

The first module was written by Mikhail Gruzdev, Thales Communications, and
Viveris Technologies. The test module was written only by Thales Communications
and Viveris Technologies.

The code property on the first module is shared among several people/entities,
because the work was done twice. This work was done by Viveris Technologies
for Thales Communications in a private fork several months ago (along with the
2nd module for testing). Then, the first module was done as a separate work by
Mikhail Gruzdev. Mikhail published his work on the ROHC mailing list. Thales
Communications and Viveris Technologies agreed to publish their private work,
so the result is a merge of everything.