Introduce the configure option --disable-deprecated-api.
This option allows a developer to test whether a program uses some deprecated
API of the ROHC library by not including the deprecated API in the library.

This option shall be used with caution. It breaks API/ABI compatibility with
previous versions of the ROHC library. Do not report bugs when using the
library built with deprecated API disabled. Do not use the option when
packaging a general-purpose version of the ROHC library.