Fix bug #1038637: RTP profile: RTP header with extension flag set not correctly handled https://bugs.launchpad.net/rohc/+bug/1038637
The first packets of a new stream must transmit the value of the RTP eXtension (X) bit.
The field is considered as STATIC, so it should not change often but it must be
transmitted at least on first packets. On those packets, the RX bit shall be set to 1
even if the TSS field (for TS_STRIDE) is set to 0. Upon field change, the compressor
shall go back to FO state because the RTP eXtension bit is a STATIC field. It then can
use either one IR, IR-DYN or UO-1-ID with extension 3, or any UOR-2* with extension 3.

Always transmitting the RTP eXtension (X) bit in the first packets of a stream makes
change the reference captures of all the non-regression tests.