Fix bug #1178441: Malformed IP packet used in code example https://bugs.launchpad.net/rohc/+bug/1178441
The example application is wrong. It should compress the IP packet with the best
profile that was enabled, ie. the IP-only profile.

That's not the case because a mistake was done when building the fake IP packet.
The Total Length field is in Host Byte Order instead of Network Byte Order. The
ROHC library therefore detects that the packet is not a valid IP packet. In
consequence it chooses the Uncompressed profile instead of the IP-only profile.