Greatly improve the performances of the W-LSB decoding algorithm for large windows and lossy streams.
The W-LSB decoding algorithm was trying all the possible values in the
W-LSB window sequentialy. This algorithm is very slow for streams with
some lost packets and large windows.

The new algorithm only tries the values in the W-LSB window that match
the received k LSB bits. Only 1 or 2 tries are required.