Fix bug #1210498: IP-ID shall be transmitted verbatim in IR and IR-DYN packets https://bugs.launchpad.net/rohc/+bug/1210498
According to RFC 4815 (https://tools.ietf.org/html/rfc4815) and discussion on
the IETF list (http://www.ietf.org/mail-archive/web/rohc/current/msg01675.html):
 - "IP-ID" means that the Identification field is transmitted encoded
   with the method "Offset IP-ID", ie. offset between SN and IP-ID and then
   byte-swapped according to the NBO flag.
 - "Identification" means that the Identification field is transmitted
   verbatim, ie. unchanged wrt the non compressed packet.

So, the ROHC library is not conform: the dynamic part of the IPv4 header in the
IR and IR-DYN packets shall contain the verbation Identification field, not the
byte-swapped value.

Problem discovered with the help of Fabrice Bellard.

THIS CHANGE MAKES THE LIBRARY INCOMPATIBLE WITH ANY PREVIOUS VERSION for IPv4
streams with IP-ID incrementing in little endian.