Add support for repair upon CRC failure as described in RFC 3095, §5.3.2.2.3.
The patch implements the 2 algorithms:
 - Correction of SN LSB wraparound as described in RFC 3095, §5.3.2.2.4.
 - Repair of incorrect SN updates as described in RFC 3095, §5.3.2.2.5.

The detection of SN LSB wraparound is not implemented exactly as described in
the RFC (this is OK for RFC if the algorithm is as robust as their):
 - According to RFC3095, §5.3.2.2.4, step c, SN wraparound is possible if the
 inter-packet interval of the current packet is at least 2^k times the nominal
 inter-packet interval (with k the number of SN bits in the current header).
 - However SN wraparound may happen sooner depending on the shift parameter p
 of the W-LSB algorithm. If p is large, the interpretation interval is shifted
 on the left: the positive part of the interpretation interval is smaller.
 Less (lost) packets are needed to cause a wraparound.
 - The 'width of the positive part of the interpretation interval' (2^k - p) is
 used instead of the 'width of the full interpretation interval' (2^k).
 - Finally, a -10% marge is taken to handle problems due to clock precision.

The detection of SN LSB wraparound requires the arrival time of every packet.
The library was already recording arrival times for statistical purposes. The
precision was not enough however. Retrieving time with high prevision highly
depends on the platform (Linux, BSD, Windows...), so it was decided to let
users retrieve arrival times. In addition, users may have a better knowledge
of the arrival time of the packet if they don't decompress ROHC packet
immediately when they are received. Last, it eases library testing...

In order to let the users give the arrival time, the rohc_compress2() and
rohc_decompress() functions were updated to take an additional argument: the
arrival time, stored as struct timespec. The patch therefore introduces the
rohc_compress3() and rohc_decompress2() functions that deprecates the previous
ones. All applications, examples, and tests were updated in consequence.

The test_lost_packet (resp. test_damaged_packet) test was also updated to check
the correct repair of packet (resp. context) after a packet was lost (resp.
damaged).