API: add new function rohc_feedback_avail_bytes() in compressor API.
A new function rohc_feedback_avail_bytes() was added to the compressor API. It
returns the number of feedback bytes that are stored in the compressor waiting
to be transmitted to the related decompressor on the other end of the feedback
channel. Only available bytes are counted. Feedback bytes that are locked
because being transmitted are not counted.

Designed with the help of Friedrich. See the discussion here:
  https://answers.launchpad.net/rohc/+question/233909