Fix an endianness problem on PowerPC with fix for bug #1210498.
Bug #1210498 is entitled "IP-ID shall be transmitted verbatim in IR and IR-DYN
packets".

On PowerPC, the IP-ID bits of the IR and IR-DYN packets shall be swapped when
decoded if the IP-ID behavior is 'increase in Little Endian'. Otherwise, they
will be swapped when written in the decompressed IP packet.