API: build an additional librohc.so to aggregate librohc_*.so
From ever the ROHC library was split into 3 .so files: librohc_common.so,
librohc_comp.so, and librohc_decomp.so. This was done so that a program
could load compression code without loading decompression code. In fact,
such programs should be quite rare, and the 3 .so files are quite small
(~650KB all once striped). So, give up and provide a librohc.so.

The new librohc.so currently contains no code, it only depends on the 3
others .so files. The other .so files will be made convenient libraries
in the 2.0.0 releases, and not installed anymore.

Update the pkg-config file rohc.pc right now. That way, developers using
pkg-config will automatically use the new library instead of the old ones.