Remove the non-standard feature "RTP bit type".
The RTP bit type uses an additionnal bit to distinguish UOR-2-TS and UOR-2-ID
packets. The bit is taken from the CRC-7 that becomes CRC-6.

The feature is removed because it is not part of any IETF RFC, and it is a
maintenance burden. And there is no known users.