TCP profile: handle context re-use with similar TCP connection
Handle correctly the compression/decompression of TCP options if a TCP
connection re-uses the context of a previous TCP connection with the
very same (src addr, src port, dst addr, dst port) tuple.

The compression/decompression of TCP options in the co_common packet
was not robust enough to handle such a case.