Introduce a new application: a ROHC fuzzer for RTP traffic
The ROHC RTP fuzzer is a fuzzing tool that compresses, then decompresses some
RTP traffic. The RTP traffic is generated with some fields changing randomly.

The ROHC RTP fuzzer application may be used the same way as the existing ROHC
fuzzer: command 'play' to run the tool, command 'replay SEED' to replay the
same run to reproduce a bug.