Fix bug #1265304: RTP decompression failure with repeated constant TS https://bugs.launchpad.net/rohc/+bug/1265304
If RTP TS is constant, it must be transmitted uncompressed. It is however sent
encoded with W-LSB. So, given enough RTP packets with the same constant RTP TS
value, 0 bit are needed by the decompressor to decode the value. As it is
forbidden to not sent at least 1 bit of TS when it is constant, so the compressor
forces the transmission of 1 bit. In the meanwhile, the TS bits were reset, so
the compressor sends bits(TS) = 0. That's wrong. The TS bit shall be extracted
again from the uncompressed TS in such a case.

Improve a little bit the robustness by sending 7 bits of TS instead of 1. The
field is sent encoded in SDVL, so 1 or 7 bits take 1 byte. Better send 7 bits!