man pages: handle parallel make when using help2man
The man pages of the apps are built with a tool named help2man. help2man runs
the app binary with --help as argument and generates a man page from the output.
For this process to work fine the app binary shall be built before help2man is
called. With parallel make (-j5 for example) this was not properly ensured. This
patch adds a make dependency on the app binaries.