W-LSB: compute the minimum number of required bits for all values in window
The ROHC decompressor shall be able to decode the W-LSB-compressed value with
any reference value in the compressor window. The compressor computed the
minimum number of required bits based on the minimum and maximum values of the
window, but it is wrong: the compressor shall compute the minimum number of
required bits based on all the values in the window.

Example with window values: { 4294961711, 4294961715, 2623, 2632 }
The minimum number of required bits was computed based on the minimum
value (2623) and the maximum value (4294961715). This is wrong because a
wraparound happened: minimum is 4294961715 and maximum is 2632.

Because it is difficult to reliabily detect a wraparound (think about
reordering), better compute the minimum number of required bits based on all
values in the window, and keep only the greatest.