Create a ROHC stats application from the ./statistics/* tests.
The tests located in the ./statistics/ subdir were n,ot very useful. The
compression statistics generated from the testing captures were not
representative of any real-world traffic.

In addition, it was difficult to generate the statistics for a network
capture that was not a test capture:
 - the ./statistics/generate_statistics binary was not enough to generate
    interesting graphs,
 - the ./statistics/generate_statistics.sh script was able to generate
   interesting graphs but it was designed to get network captures from a
   fixed location (non-regression tests).

The new ROHC stats application is not run as a autotools test anymore. It
may be run against any network capture easily. It doesn't generate graphs
but the ./app/stats/rohc_stats.sh shell script is able to do so for one
single capture.