TCP profile: dynamic chain decoding: do not write a length field for EOL option
If padding bytes are present at the very end of list of TCP options after the
EOL option, the dynamic chain contains a list item of type EOL with a 1-byte
field that represents the number of padding bytes. So, the EOL option (and the
following padding) is not always 1-byte long, it could be larger.

If it is larger, we should not write the length in the 2nd byte: EOL is a short
option not a long one.