non-regression: add scripts to download and test external/personal captures.
Sometimes we have a network capture that causes a trouble to the ROHC library.
We fix the problem and we would like the problem not to come back later. The
best way to achieve this is to put the network capture in the non-regression
tests.

However, it might happen that the capture contains sensitive and personal data,
so uploading it in a public location is not an option. There is now an
alternative:
 1/ put the capture in:
     test/non_regression/other_inputs/personal_inputs/
 2/ then run:
     $ make -C test/non_regression/other_inputs/ check_personal

However, it might happen that the capture is under copyright and cannot be
published with ROHC. The capture might also be huge (> 1 GB). There is now an
alternative:
 1/ download the captures:
     $ make -C test/non_regression/other_inputs/ download
 2/ then run:
     $ make -C test/non_regression/other_inputs/ check_external