use the Uncompressed profile if IPv4 checksum is incorrect
If the IPv4 checksum is incorrect (both for outer or inner IP header), the ROHC
compression/decompression won't be able to transmit the packet unchanged byte
per byte. The IPv4 checksum will be computed correctly at decompressor and it
won't match the original one.

To avoid this, verify the IPv4 checksums at compressor and avoid to use the
IP-based profiles if at least one of them is incorrect. Use the Uncompressed
profile instead.

This behaviour is the safest way to go. However, verifying the checksums might
be time consuming and might degrade the CPU performances, so add a compressor
feature to disable that verification. The feature is named:
	ROHC_COMP_FEATURE_NO_IP_CHECKSUMS
It may be set with API function rohc_comp_set_features().