add shortcuts rohc_comp_warn() and rohc_comp_list_warn() for rohc_warning()
The macro rohc_comp_warn() may be used with a compression context as first
parameter to log a warning trace with the context profile.

The macro rohc_comp_list_warn() may be used with a list compression context
as first parameter to log a warning trace with the profile associated to the
list compression context.