add shortcuts rohc_decomp_warn() and rd_list_warn() for rohc_warning()
The macro rohc_decomp_warn() may be used with a decompression context as first
parameter to log a warning trace with the context profile.

The macro rd_list_warn() may be used with a list decompression context as first
parameter to log a warning trace with the profile associated to the list
decompression context.