Add a network buffer implementation
A network buffer may represent one uncompressed packet, one ROHC packet, or one
ROHC feedback.

The network buffer does not contain the packet data itself. It only has a
pointer on it. This is designed this way for performance reasons: no copy
required to initialize a network buffer, the struct is small and may be passed
as copy to function.

The network buffer is able to keep some free space at its beginning. The
unused space at the beginning of the buffer may be used to prepend a network
header at the very end of the packet handling.

The beginning of the network buffer may also be shifted forward and backward
with the rohc_buf_shift function. This is useful when parsing a network packet
(once bytes are read, shift them forward) for example.

The network buffer may be initialized manually by filling all the fields of the
struct or with the helper functions rohc_buf_init_empty or rohc_buf_init_full.