Use the new network buffer for compression API: introduce rohc_compress4()
the rohc_compress4() function provides the same features as rohc_compress3()
but it is based on the new network buffer (struct rohc_buf). The network
buffer gathers the packet buffer, the packet length and the packet timestamp.
Both for uncompressed and compressed packets.

So rohc_compress3() signature with 7 parameters:

	int rohc_compress3(struct rohc_comp *const comp,
	                   const struct rohc_ts arrival_time,
	                   const unsigned char *const uncomp_packet,
	                   const size_t uncomp_packet_len,
	                   unsigned char *const rohc_packet,
	                   const size_t rohc_packet_max_len,
	                   size_t *const rohc_packet_len);

is reduced to rohc_compress4() signature with only 3 parameters:

	int rohc_compress4(struct rohc_comp *const comp,
	                   const struct rohc_buf uncomp_packet,
	                   struct rohc_buf *const rohc_packet);

The rohc_compress4() signature is much more simpler and clearer than the one
of rohc_compress3(). That's a great improvement!

More to come on other API functions with too many parameters.