Use the new network buffer for decompression API: introduce rohc_decompress3()
The rohc_decompress3() function provides the same features as rohc_decompress2()
but it is based on the new network buffer (struct rohc_buf). The network buffer
gathers the packet buffer, the packet length and the packet timestamp. Both for
compressed and uncompressed packets.

So rohc_decompress2() signature with 7 parameters:

	int rohc_decompress2(struct rohc_decomp *const decomp,
	                     const struct rohc_ts arrival_time,
	                     const unsigned char *const rohc_packet,
	                     const size_t rohc_packet_len,
	                     unsigned char *const uncomp_packet,
	                     const size_t uncomp_packet_max_len,
	                     size_t *const uncomp_packet_len);

is reduced to rohc_decompress3() signature with only 3 parameters:

	int rohc_decompress3(struct rohc_decomp *const decomp,
	                     const struct rohc_buf rohc_packet,
	                     struct rohc_buf *const uncomp_packet);

The rohc_decompress3() signature is much more simpler and clearer than the one
of rohc_decompress2(). That's a great improvement!